+++ EssentialFeedTests/Feed Cache/LoadFeedImageDataFromCacheUseCaseTests.swift
@@ -1,5 +1,5 @@
 //
-//  LocalFeedImageDataFromCacheUseCaseTests.swift
+//  LoadFeedImageDataFromCacheUseCaseTests.swift
 //  EssentialFeedTests
 //
 //  Created by Jessie Elliott on 5/9/24.
@@ -8,22 +8,11 @@
 import XCTest
 import EssentialFeed
 
-class LocalFeedImageDataFromCacheUseCaseTests: XCTestCase {
-    
+class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
     func test_init_doesNotMessageStoreUponCreation() {
         let (_, store) = makeSUT()
         
         XCTAssertTrue(store.receivedMessages.isEmpty)
-    }
-    
-    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
-        let (sut, store) = makeSUT()
-        let url = anyURL()
-        let data = anyData()
-        
-        sut.save(data, for: url) { _ in }
-        
-        XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
     }
     
     func test_loadImageDataFromURL_requestsStoredDataForURL() {
